+++ web/documentserver-example/nodejs/app.js
@@ -27,6 +27,7 @@ const jwt = require('jsonwebtoken');
 const config = require('config');
 const mime = require('mime');
 const urllib = require('urllib');
+const urlModule = require("url");
 const { emitWarning } = require('process');
 const DocManager = require('./helpers/docManager');
 const documentService = require('./helpers/documentService');
@@ -495,6 +496,20 @@ app.post('/reference', (req, res) => { // define a handler for renaming file
                 && req.DocManager.existsSync(req.DocManager.storagePath(fileKey.fileName, userAddress))) {
         ({ fileName } = fileKey);
       }
+    }
+  }
+
+  if (!fileName && !!req.body.link) {
+    if (req.body.link.indexOf(req.DocManager.curUserHostAddress()) != -1) {
+      result({ error: 'You do not have access to this site' });
+      return;
+    }
+
+    let urlObj = urlModule.parse(req.body.link, true);
+    fileName = urlObj.query.fileName;
+    if (!req.DocManager.existsSync(req.DocManager.storagePath(fileName, userAddress))) {
+      result({ error: 'File is not exist' });
+      return;
     }
   }